++ add Run Demo class with all 11-17 cases to show implement Run Demo into Program.cs

+++ APBD_Wypozyczalnia_Proj/Services/RentalService.cs
@@ -22,7 +22,7 @@ public class RentalService
         return rental;
     }
 
-    public void ReturnEquipment(int rentalId)
+    public void ReturnEquipment(long rentalId)
     {
         var rental = _rentals.FirstOrDefault(r => r.Id == rentalId);
         if (rental == null || rental.IsReturned)
